Create db.models.tags
