+++ db/models/tags.py
@@ -1,0 +1,14 @@
+from db.database import Base
+from sqlalchemy import Column, String, Integer, TIMESTAMP, text
+from sqlalchemy.orm import relationship
+
+
+class Tag(Base):
+    __tablename__ = 'tags'
+    id = Column(Integer, primary_key=True, index=True)
+    name = Column(String)
+    created_at = Column(TIMESTAMP(timezone=True),
+                        nullable=False, server_default=text('now()'))
+    updated_at = Column(TIMESTAMP(timezone=True),
+                        nullable=False, server_default=text('now()'))
+    users = relationship('User', secondary='users_tags', back_populates='tags')
