+++ db/models/users.py
@@ -37,17 +37,6 @@ class User(Base):
     stack = relationship("Stack", back_populates='users')  
 
 
-class Tag(Base):
-    __tablename__ = 'tags'
-    id = Column(Integer, primary_key=True, index=True)
-    name = Column(String)
-    created_at = Column(TIMESTAMP(timezone=True),
-                        nullable=False, server_default=text('now()'))
-    updated_at = Column(TIMESTAMP(timezone=True),
-                        nullable=False, server_default=text('now()'))
-    users = relationship('User', secondary='users_tags', back_populates='tags')
-
-
 class Feed(Base):
     __tablename__ = 'feeds'
     id = Column(Integer, primary_key=True, nullable=False)
